abusehelper.bots.archivebot: Fix for path handling
 * Refactor all ".compress-********" path handling into one location in
the file, as disrepancies in handling such paths seems to be a common
source of problems. Add tests.

 * Add archivebot._unique_writable_file that should be less prone to
race conditions. The previous version of the corresponding code first
created the filename, then checked whether such file already exists,
and then opened the file. The new version creates the filename with the
O_CREAT and O_EXCL flags set and thus fails when the file already
exists.