abusehelper.core.utils.fetch_url: Check server certificates by default
Remove fetch_url support for custom urllib2 openers with the 'opener'
keyword argument. Trying to pass in a custom opener now raises a TypeError.

Add arguments 'cookies', 'auth', 'cert', 'verify' and 'proxies', in the
footsteps of the Requests library. However the functionality here is very
limited (and probably incompatible) when compared to Requests.

Signed-off-by: Ossi Herrala <oherrala@gmail.com>