Make starting new session default behaviour
RFC3501 says:

The unique identifier of a message MUST NOT change during the session,
and SHOULD NOT change between sessions.

Thus it is perfectly acceptable that UID changes between sessions, so
remove the option for starting from fresh state when connection is lost
and instead make it the default behaviour.